* Few corrections for GCC, and now 9box demo nearly works
* Few modifications for Linux.
* Added keywords to main EMSCENE.lang
* Integrated existing documents.
* Modernisation of miniline code.
* Modifications of stream management, for writing and reading objects.
* Modernizing the code.
* Manager now creates and stores std::shared_ptr< TObj >.
* owned_by< T > needs a reference instead of a pointer, in ctor.
* Events have been reviewed a bit, to avoid storing a void *, needs some more work though.
* Removed Condtion, Mutex and RecursiveMutex typedefs, and related macros.
* Cleaned up Macros.h, removed unused ones (that is, all of them).
* Lots of updates.
* Miniline server relaunches successfully.
* Fixed most warnings, compiling in x64.
* Removed all namespace macros from Troll.
* Removed pointers to Scene into Project.
* Implemented zip folder using wxWidgets.
* Moved teproj writing to ProjectFileWriter new class.
* Moved muse writing to MuseFileWriter new class.
* Project creation and modifications now use wxPropertyGrids.
* Removed generated files from data folder.
* Added kazu splash in every kazu game.
* Removed tree item management from Scene and Project.
* Added ProjectFrame class, which allows the project unloading easily.
* Replacing properties.
* Properties are replaced.
* Applied AUI theme.
* Updated CMake submodule

+++ source/Core/GeneralLib/Src/FastMath.h
@@ -38,7 +38,7 @@ namespace General
 			return p_number * y;
 		}
 
-		inline unsigned int NextPowOfTwo( unsigned int x )
+		inline uint32_t NextPowOfTwo( uint32_t x )
 		{
 			-- x;
 			x |= ( x >> 1 );
@@ -58,14 +58,14 @@ namespace General
 
 		inline double FastAbs( double p_float )
 		{
-			long long * j = reinterpret_cast <long long *>( & p_float );
+			int64_t * j = reinterpret_cast <int64_t *>( & p_float );
 			( * j ) = ( * j ) & 0x7FFFFFFFFFFFFFFFLL;
 			return ( * ( reinterpret_cast <double *>( j ) ) );
 		}
 
-		inline unsigned int uabs( int p_value )
+		inline uint32_t uabs( int p_value )
 		{
-			return static_cast <unsigned int>( p_value & 0x7FFFFFFF );
+			return static_cast <uint32_t>( p_value & 0x7FFFFFFF );
 		}
 
 		template< typename T >
